Move fuzz_select over to select in interactive
